// src/lib/api-types.ts
export interface components {
    schemas: {
// ...
        CampaignLeadResponse: {
            /** Id */
            id: string;
            /** Company Campaign Id */
            company_campaign_id: string;
            /** External Lead Id */
            external_lead_id: string;
            /** Email */
            email?: string | null;
            /** First Name */
            first_name?: string | null;
            /** Last Name */
            last_name?: string | null;
            /** Company Name */
            company_name?: string | null;
            /** Title */
            title?: string | null;
            /**
             * Status
             * @enum {string}
             */
            status: "active" | "paused" | "unsubscribed" | "replied" | "bounced" | "pending" | "contacted" | "connected" | "not_interested" | "unknown";
            /** Category */
            category?: string | null;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}